refactor: models to modular pattern
