+++ app/Modules/Customer/Models/Customer.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace App\Models;
+namespace App\Modules\Customer\Models;
 
+use App\Modules\Order\Models\Order;
+use App\Modules\User\Models\User;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
@@ -9,18 +11,18 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Order extends Model
+class Customer extends Model
 {
     use HasFactory, softDeletes;
 
-    public function customer(): BelongsTo
+    public function orders(): HasMany
     {
-        return $this->belongsTo(Customer::class);
+        return $this->hasMany(Order::class);
     }
 
-    public function order_items(): HasMany
+    public function user(): BelongsTo
     {
-        return $this->hasMany(OrderItem::class);
+        return $this->belongsTo(User::class);
     }
 
     public function scopeSearch(Builder $query, ?string $search): Builder
