+++ app/Modules/Inventory/Models/Inventory.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Models;
+namespace App\Modules\Inventory\Models;
 
+use App\Modules\Product\Models\Product;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
