+++ app/Modules/Log/Models/Log.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Models;
+namespace App\Modules\Log\Models;
 
+use App\Modules\User\Models\User;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
