+++ app/Modules/Message/Models/Contact.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Models;
+namespace App\Modules\Message\Models;
 
+use App\Modules\User\Models\User;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
