+++ app/Modules/Order/Models/Order.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Models;
+namespace App\Modules\Order\Models;
 
+use App\Modules\Customer\Models\Customer;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
@@ -9,18 +10,18 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Customer extends Model
+class Order extends Model
 {
     use HasFactory, softDeletes;
 
-    public function orders(): HasMany
+    public function customer(): BelongsTo
     {
-        return $this->hasMany(Order::class);
+        return $this->belongsTo(Customer::class);
     }
 
-    public function user(): BelongsTo
+    public function order_items(): HasMany
     {
-        return $this->belongsTo(User::class);
+        return $this->hasMany(OrderItem::class);
     }
 
     public function scopeSearch(Builder $query, ?string $search): Builder
