+++ app/Modules/Order/Models/OrderItem.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Models;
+namespace App\Modules\Order\Models;
 
+use App\Modules\Product\Models\Product;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
