+++ app/Modules/Product/Models/Product.php
@@ -1,13 +1,16 @@
 <?php
 
-namespace App\Models;
+namespace App\Modules\Product\Models;
 
+use App\Modules\Category\Models\Category;
+use App\Modules\Inventory\Models\Inventory;
+use App\Modules\Order\Models\OrderItem;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Product extends Model
 {
